Sync upstream and support pinned K2-Horizon MoVA 6.50bpw

+++ exllamav3/exllamav3_ext/libtorch/attention.h
@@ -59,7 +59,7 @@ struct BC_Attention
     c10::optional<at::Tensor> kv_ptrs_trellis;
     c10::optional<at::Tensor> kv_ptrs_suh;
     c10::optional<at::Tensor> kv_ptrs_svh;
-    int kv_K;
+    float kv_K;
     bool kv_mcg;
     bool kv_mul1;
     std::shared_ptr<BC_LinearEXL3> o_proj;
@@ -79,7 +79,7 @@ struct BC_Attention
     c10::optional<at::Tensor> qg_ptrs_trellis;
     c10::optional<at::Tensor> qg_ptrs_suh;
     c10::optional<at::Tensor> qg_ptrs_svh;
-    int qg_K;
+    float qg_K;
     bool qg_mcg;
     bool qg_mul1;
 
@@ -91,7 +91,7 @@ struct BC_Attention
     c10::optional<at::Tensor> qkv_ptrs_suh;      // per source
     c10::optional<at::Tensor> qkv_ptrs_svh;
     c10::optional<at::Tensor> qkv_meta;
-    int qkv_K;
+    float qkv_K;
     bool qkv_mcg;
     bool qkv_mul1;
     at::Tensor qkv_size_n, qkv_n_stride, qkv_had_src;   // device int32, per slice
@@ -205,7 +205,7 @@ struct BC_Attention
         c10::optional<at::Tensor> kv_ptrs_trellis,
         c10::optional<at::Tensor> kv_ptrs_suh,
         c10::optional<at::Tensor> kv_ptrs_svh,
-        int kv_K,
+        float kv_K,
         bool kv_mcg,
         bool kv_mul1,
         std::shared_ptr<BC_LinearEXL3> o_proj,
@@ -217,14 +217,14 @@ struct BC_Attention
         c10::optional<at::Tensor> qg_ptrs_trellis,
         c10::optional<at::Tensor> qg_ptrs_suh,
         c10::optional<at::Tensor> qg_ptrs_svh,
-        int qg_K,
+        float qg_K,
         bool qg_mcg,
         bool qg_mul1,
         c10::optional<at::Tensor> qkv_ptrs_trellis,
         c10::optional<at::Tensor> qkv_ptrs_suh,
         c10::optional<at::Tensor> qkv_ptrs_svh,
         c10::optional<at::Tensor> qkv_meta,
-        int qkv_K,
+        float qkv_K,
         bool qkv_mcg,
         bool qkv_mul1,
         c10::optional<at::Tensor> q_norm,
