+++ exllamav3/exllamav3_ext/quant/reconstruct.cuh
@@ -6,7 +6,7 @@ void reconstruct
 (
     at::Tensor unpacked,
     at::Tensor packed,
-    int K,
+    float K,
     bool mcg,
     bool mul1
 );
@@ -15,7 +15,7 @@ void reconstruct_slice
 (
     at::Tensor unpacked,
     at::Tensor packed,
-    int K,
+    float K,
     bool mcg,
     bool mul1,
     int64_t n_offset
@@ -27,7 +27,7 @@ void reconstruct_had_slice
     at::Tensor packed,
     at::Tensor suh,
     at::Tensor svh,
-    int K,
+    float K,
     bool mcg,
     bool mul1,
     int64_t n_offset
@@ -39,7 +39,7 @@ void reconstruct_had_batch
     at::Tensor packed_ptrs,
     at::Tensor suh_ptrs,
     at::Tensor svh_ptrs,
-    int K,
+    float K,
     bool mcg,
     bool mul1
 );
@@ -48,7 +48,7 @@ void reconstruct_batch
 (
     at::Tensor unpacked,
     at::Tensor packed_ptrs,
-    int K,
+    float K,
     bool mcg,
     bool mul1
 );
